// ams-fe/src/app/model/requestDto.ts
/**
 * Demo REST API
 * User Management API.
 *
 * OpenAPI spec version: API v1.0.0
 * Contact: nhtoi@tma.com.vn
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface RequestDto { 
    assetIds?: number;
    reason?: string;
    requestType?: string;
    status?: string;
    updateAt?: Date;
    userIdApprove?: Array<number>;
    userIdReceive?: Array<number>;
    userIdSend?: Array<number>;
}